m7

+++ src/main/java/com/driver/services/BlogService.java
@@ -37,7 +37,9 @@ public class BlogService {
         //create a blog at the current time
 
         Blog blog = new Blog(title,content);
+        blog.setPubDate(new Date());
         User user = userRepository1.findById(userId).get();
+        blog.setUser(user);
         List<Blog> blogList = user.getBlogList();
         blogList.add(blog);
         user.setBlogList(blogList);
@@ -58,10 +60,13 @@ public class BlogService {
 
     public void addImage(Integer blogId, String description, String dimensions){
         //add an image to the blog after creating it
-        Image image = new Image(description,dimensions);
         Blog blog = blogRepository1.findById(blogId).get();
+        Image image = imageService1.createAndReturn(blog,description,dimensions);
+
         image.setBlog(blog);
         List<Image> imageList= blog.getImageList();
+        if(imageList==null)
+            imageList = new ArrayList<>();
         imageList.add(image);
         blog.setImageList(imageList);
         blogRepository1.save(blog);
@@ -70,6 +75,8 @@ public class BlogService {
 
     public void deleteBlog(int blogId){
         //delete blog and corresponding images
+        if(blogRepository1.findById(blogId).get()==null)
+            return;
         blogRepository1.deleteById(blogId);
     }
 }
